% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{myfunction}
\alias{myfunction}
\title{Title Will give a positive version of the number}
\usage{
myfunction(x)
}
\arguments{
\item{x}{Takes the number as an input}
}
\value{
a positive side of the number
}
\description{
Title Will give a positive version of the number
}
